requirements.txt
Flask==1.1.1
PyMySQL==0.9.3
python-dotenv==0.10.3
